added GuildChannel class
